Rename file line-chart.vue to LineChart.vue
